refactor: move state into useSafeMessage hook
